refactor imports and command descriptions in cogs; gitignore update
